extract(module)
Extracts the Range module from the Build
